LPS-115840 Move methods used to retrieve BlogsEntry fieldValues to an implementation of InfoItemFieldValuesProvider
Extract field definitions into the new BlogsEntryInfoItemFields interface to avoid duplication
