+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/info/item/provider/BlogsEntryInfoItemFormProvider.java
@@ -18,39 +18,15 @@ import com.liferay.asset.display.page.portlet.AssetDisplayPageFriendlyURLProvide
 import com.liferay.asset.info.item.provider.AssetEntryInfoItemFieldSetProvider;
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.web.internal.info.item.BlogsEntryInfoItemFields;
 import com.liferay.expando.info.item.provider.ExpandoInfoItemFieldSetProvider;
-import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldSetEntry;
-import com.liferay.info.field.InfoFieldValue;
-import com.liferay.info.field.type.ImageInfoFieldType;
-import com.liferay.info.field.type.TextInfoFieldType;
-import com.liferay.info.field.type.URLInfoFieldType;
 import com.liferay.info.form.InfoForm;
-import com.liferay.info.item.InfoItemClassPKReference;
-import com.liferay.info.item.InfoItemFieldValues;
-import com.liferay.info.item.NoSuchInfoItemException;
 import com.liferay.info.item.field.reader.InfoItemFieldReaderFieldSetProvider;
 import com.liferay.info.item.provider.InfoItemFormProvider;
-import com.liferay.info.localized.InfoLocalizedValue;
-import com.liferay.info.type.WebImage;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
-import com.liferay.portal.kernel.util.LocaleThreadLocal;
 
-import java.text.Format;
-
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
 
 import org.osgi.framework.Constants;
 import org.osgi.service.component.annotations.Component;
@@ -88,157 +64,19 @@ public class BlogsEntryInfoItemFormProvider
 		return infoForm;
 	}
 
-	@Override
-	public InfoItemFieldValues getInfoFormValues(BlogsEntry blogsEntry) {
-		InfoItemFieldValues infoItemFieldValues = new InfoItemFieldValues(
-			new InfoItemClassPKReference(
-				BlogsEntry.class.getName(), blogsEntry.getEntryId()));
-
-		infoItemFieldValues.addAll(_getBlogsEntryInfoFieldValues(blogsEntry));
-
-		try {
-			infoItemFieldValues.addAll(
-				_assetEntryInfoItemFieldSetProvider.getInfoFieldValues(
-					BlogsEntry.class.getName(), blogsEntry.getEntryId()));
-		}
-		catch (NoSuchInfoItemException noSuchInfoItemException) {
-			throw new RuntimeException(
-				"Caught unexpected exception", noSuchInfoItemException);
-		}
-
-		infoItemFieldValues.addAll(
-			_expandoInfoItemFieldSetProvider.getInfoFieldValues(
-				BlogsEntry.class.getName(), blogsEntry));
-		infoItemFieldValues.addAll(
-			_infoItemFieldReaderFieldSetProvider.getInfoFieldValues(
-				BlogsEntry.class.getName(), blogsEntry));
-
-		return infoItemFieldValues;
-	}
-
 	private Collection<InfoFieldSetEntry> _getBlogsEntryInfoFieldSetEntries() {
 		return Arrays.asList(
-			_titleInfoField, _subtitleInfoField, _descriptionInfoField,
-			_smallImageInfoField, _coverImageInfoField,
-			_coverImageCaptionInfoField, _authorNameInfoField,
-			_authorProfileImageInfoField, _publishDateInfoField,
-			_displayPageUrlInfoField, _contentInfoField);
-	}
-
-	private List<InfoFieldValue<Object>> _getBlogsEntryInfoFieldValues(
-		BlogsEntry blogsEntry) {
-
-		List<InfoFieldValue<Object>> blogsEntryFieldValues = new ArrayList<>();
-
-		ThemeDisplay themeDisplay = _getThemeDisplay();
-
-		try {
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(_titleInfoField, blogsEntry.getTitle()));
-
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(
-					_subtitleInfoField, blogsEntry.getSubtitle()));
-
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(
-					_descriptionInfoField, blogsEntry.getDescription()));
-
-			if (themeDisplay != null) {
-				WebImage smallWebImage = new WebImage(
-					blogsEntry.getSmallImageURL(themeDisplay));
-
-				smallWebImage.setAlt(blogsEntry.getSmallImageAlt());
-
-				blogsEntryFieldValues.add(
-					new InfoFieldValue<>(_smallImageInfoField, smallWebImage));
-
-				WebImage coverWebImage = new WebImage(
-					blogsEntry.getCoverImageURL(themeDisplay));
-
-				coverWebImage.setAlt(blogsEntry.getCoverImageAlt());
-
-				blogsEntryFieldValues.add(
-					new InfoFieldValue<>(_coverImageInfoField, coverWebImage));
-			}
-
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(
-					_coverImageCaptionInfoField,
-					blogsEntry.getCoverImageCaption()));
-
-			User user = _userLocalService.fetchUser(blogsEntry.getUserId());
-
-			if (user != null) {
-				blogsEntryFieldValues.add(
-					new InfoFieldValue<>(
-						_authorNameInfoField, user.getFullName()));
-
-				if (themeDisplay != null) {
-					WebImage webImage = new WebImage(
-						user.getPortraitURL(themeDisplay));
-
-					webImage.setAlt(user.getFullName());
-
-					blogsEntryFieldValues.add(
-						new InfoFieldValue<>(
-							_authorProfileImageInfoField, webImage));
-				}
-			}
-
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(
-					_publishDateInfoField,
-					_getDateValue(blogsEntry.getDisplayDate())));
-
-			if (themeDisplay != null) {
-				blogsEntryFieldValues.add(
-					new InfoFieldValue<>(
-						_displayPageUrlInfoField,
-						_getDisplayPageURL(blogsEntry)));
-			}
-
-			blogsEntryFieldValues.add(
-				new InfoFieldValue<>(
-					_contentInfoField, blogsEntry.getContent()));
-
-			return blogsEntryFieldValues;
-		}
-		catch (PortalException portalException) {
-			throw new RuntimeException(portalException);
-		}
-	}
-
-	private String _getDateValue(Date date) {
-		if (date == null) {
-			return StringPool.BLANK;
-		}
-
-		Locale locale = LocaleThreadLocal.getThemeDisplayLocale();
-
-		Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(
-			locale);
-
-		return dateFormatDateTime.format(date);
-	}
-
-	private String _getDisplayPageURL(BlogsEntry blogsEntry)
-		throws PortalException {
-
-		return _assetDisplayPageFriendlyURLProvider.getFriendlyURL(
-			BlogsEntry.class.getName(), blogsEntry.getEntryId(),
-			_getThemeDisplay());
-	}
-
-	private ThemeDisplay _getThemeDisplay() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getThemeDisplay();
-		}
-
-		return null;
+			BlogsEntryInfoItemFields.titleInfoField,
+			BlogsEntryInfoItemFields.subtitleInfoField,
+			BlogsEntryInfoItemFields.descriptionInfoField,
+			BlogsEntryInfoItemFields.smallImageInfoField,
+			BlogsEntryInfoItemFields.coverImageInfoField,
+			BlogsEntryInfoItemFields.coverImageCaptionInfoField,
+			BlogsEntryInfoItemFields.authorNameInfoField,
+			BlogsEntryInfoItemFields.authorProfileImageInfoField,
+			BlogsEntryInfoItemFields.publishDateInfoField,
+			BlogsEntryInfoItemFields.displayPageUrlInfoField,
+			BlogsEntryInfoItemFields.contentInfoField);
 	}
 
 	@Reference
@@ -249,53 +87,11 @@ public class BlogsEntryInfoItemFormProvider
 	private AssetEntryInfoItemFieldSetProvider
 		_assetEntryInfoItemFieldSetProvider;
 
-	private final InfoField _authorNameInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "author-name"), "authorName");
-	private final InfoField _authorProfileImageInfoField = new InfoField(
-		ImageInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "author-profile-image"),
-		"authorProfileImage");
-	private final InfoField _contentInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "content"), "content");
-	private final InfoField _coverImageCaptionInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "cover-image-caption"),
-		"coverImageCaption");
-	private final InfoField _coverImageInfoField = new InfoField(
-		ImageInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "cover-image"), "coverImage");
-	private final InfoField _descriptionInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "description"), "description");
-	private final InfoField _displayPageUrlInfoField = new InfoField(
-		URLInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(
-			"com.liferay.asset.info.display.impl", "display-page-url"),
-		"displayPageURL");
-
 	@Reference
 	private ExpandoInfoItemFieldSetProvider _expandoInfoItemFieldSetProvider;
 
 	@Reference
 	private InfoItemFieldReaderFieldSetProvider
 		_infoItemFieldReaderFieldSetProvider;
-
-	private final InfoField _publishDateInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "publish-date"), "publishDate");
-	private final InfoField _smallImageInfoField = new InfoField(
-		ImageInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "small-image"), "smallImage");
-	private final InfoField _subtitleInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "subtitle"), "subtitle");
-	private final InfoField _titleInfoField = new InfoField(
-		TextInfoFieldType.INSTANCE,
-		InfoLocalizedValue.localize(getClass(), "title"), "title");
-
-	@Reference
-	private UserLocalService _userLocalService;
 
 }
